Move Database, Property & RichText into separate files :truck:
