+++ Src/Notion.Client/Models/Database/Properties/CheckboxProperty.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace Notion.Client
+{
+    public class CheckboxProperty : Property
+    {
+        public override PropertyType Type => PropertyType.Checkbox;
+        public Dictionary<string, object> Checkbox { get; set; }
+    }
+}
