+++ Src/Notion.Client/Models/Database/Properties/DateProperty.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace Notion.Client
+{
+    public class DateProperty : Property
+    {
+        public override PropertyType Type => PropertyType.MultiSelect;
+        public Dictionary<string, object> Date { get; set; }
+    }
+}
