+++ Src/Notion.Client/Models/Database/Properties/EmailProperty.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace Notion.Client
+{
+    public class EmailProperty : Property
+    {
+        public override PropertyType Type => PropertyType.Email;
+        public Dictionary<string, object> Email { get; set; }
+    }
+}
