+++ Src/Notion.Client/Models/Database/Properties/FileProperty.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace Notion.Client
+{
+    public class FileProperty : Property
+    {
+        public override PropertyType Type => PropertyType.File;
+        public Dictionary<string, object> File { get; set; }
+    }
+}
