+++ Src/Notion.Client/Models/Database/Properties/FormulaProperty.cs
@@ -1,0 +1,14 @@
+﻿namespace Notion.Client
+{
+    public class FormulaProperty : Property
+    {
+        public override PropertyType Type => PropertyType.Formula;
+
+        public Formula Formula { get; set; }
+    }
+
+    public class Formula
+    {
+        public string Expression { get; set; }
+    }
+}
