+++ Src/Notion.Client/Models/Database/Properties/LastEditedByProperty.cs
@@ -1,0 +1,13 @@
+﻿using Newtonsoft.Json;
+using System.Collections.Generic;
+
+namespace Notion.Client
+{
+    public class LastEditedByProperty : Property
+    {
+        public override PropertyType Type => PropertyType.LastEditedBy;
+
+        [JsonProperty("last_edited_by")]
+        public Dictionary<string, object> LastEditedBy { get; set; }
+    }
+}
