+++ Src/Notion.Client/Models/Database/Properties/PeopleProperty.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace Notion.Client
+{
+    public class PeopleProperty : Property
+    {
+        public override PropertyType Type => PropertyType.People;
+        public Dictionary<string, object> People { get; set; }
+    }
+}
