+++ Src/Notion.Client/Models/Database/Properties/PhoneNumberProperty.cs
@@ -1,0 +1,13 @@
+﻿using Newtonsoft.Json;
+using System.Collections.Generic;
+
+namespace Notion.Client
+{
+    public class PhoneNumberProperty : Property
+    {
+        public override PropertyType Type => PropertyType.PhoneNumber;
+
+        [JsonProperty("phone_number")]
+        public Dictionary<string, object> PhoneNumber { get; set; }
+    }
+}
