+++ Src/Notion.Client/Models/Database/Properties/TitleProperty.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace Notion.Client
+{
+    public class TitleProperty : Property
+    {
+        public override PropertyType Type => PropertyType.Title;
+        public Dictionary<string, object> Title { get; set; }
+    }
+}
