+++ Src/Notion.Client/Models/Database/Properties/UrlProperty.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace Notion.Client
+{
+    public class UrlProperty : Property
+    {
+        public override PropertyType Type => PropertyType.Url;
+        public Dictionary<string, object> Url { get; set; }
+    }
+}
