+++ Src/Notion.Client/Models/Database/RichText/RichTextType.cs
@@ -1,0 +1,19 @@
+﻿using System.Runtime.Serialization;
+
+namespace Notion.Client
+{
+    public enum RichTextType
+    {
+        [EnumMember(Value = null)]
+        Unknown,
+
+        [EnumMember(Value = "text")]
+        Text,
+
+        [EnumMember(Value = "mention")]
+        Mention,
+
+        [EnumMember(Value = "equation")]
+        Equation
+    }
+}
